collections working, ish.
Caveats are that collections don't work very well with UnityEngine.Object, as you have to use cloning with them. As a workaround, you can attempt to use POCOs, but then you have issues with binding buttons to things. A workaround is provided in the form of ICommand, but its done weird - you have to use the Command method on the DataContext, and supply the property name to the ICommand. DataContext will list all the properties that return ICommand objects.