changed how unityevents are saved/referenced, commandbindings working
as we should, the unity events are now referenced via the actual field name, and not by some property (As the property is merely a getter for scripting access).  Command bindings working.  A test build was also made from this version, and things are working in it (meaning the PostProcessScene tasks are working)