Merged PR 2906910: Fix a slowdown when searching for the type in the DataContext.
Fix a slowdown when searching for the type in the DataContext.  The previous code would result in the search query being executed every frame that the list was displayed.  The query takes on the order of 500-800ms to execute for a small project, so the Editor would appear softlocked until you managed to select one of the types.

Now we force the query to execute exactly once.

Also add/clarify labels for things.