Merged PR 3277304: Cleanup MRW specific comments in the repo
**Change Description**
Removed MRW specific comments that were added earlier to the repo. This is in preparation for pushing these changes back to the open source repo. This is the first step for preparing for the port.

Also added ifdefstatements for supporting lower .net back-end and unity 2017.

**Testing**
Compiled against unity 2017, unity 2018 and unity 5.5

Related work items: #21241895, #21672021