Fix compilation error (plus a couple other minor things) (#29)
The previous checkin introduced a compiler error. In TypeSuggestionProvider, the new Where clause needs to be above the Select clause since it is filtering on Type (not on TypeSuggestion).

I also noticed that Unity event handlers were being added when the BindingMode is OneTime which doesn't make any sense. I added some extension methods on BindingMode to help prevent this kind of mistake.

While fixing this, it also occurred to me that there is no point showing the Event dropdown in the ProeprtyBinding inspector if values never flow from the target back to the source, so now the inspector conditionally draws the event dropdown.